Refactor: delete static call #5
