+++ src/main/java/org/certis/studyplatform/auth/domain/model/vo/AccountNumberVo.java
@@ -3,14 +3,12 @@ package org.certis.studyplatform.auth.domain.model.vo;
 import org.certis.studyplatform.exception.DomainException;
 import org.certis.studyplatform.exception.ExceptionStatus;
 
-import java.util.Objects;
 
-import static org.certis.studyplatform.exception.ExceptionStatus.AUTH_DOMAIN_INVALID_ACCOUNT_NUMBER_LENGTH;
 
 public record AccountNumberVo(String accountNumber) {
     public AccountNumberVo{
         if(accountNumber.length()<6||accountNumber.length()>20){
-            throw new DomainException(AUTH_DOMAIN_INVALID_ACCOUNT_NUMBER_LENGTH);
+            throw new DomainException(ExceptionStatus.AUTH_DOMAIN_INVALID_ACCOUNT_NUMBER_LENGTH);
         }
     }
