+++ src/main/java/org/certis/studyplatform/shared/config/SecurityConfig.java
@@ -1,8 +1,8 @@
 package org.certis.studyplatform.shared.config;
 
 import lombok.RequiredArgsConstructor;
-import org.certis.studyplatform.auth.infrastructure.security.JwtAuthenticationFilter;
-import org.certis.studyplatform.auth.infrastructure.security.JwtTokenProvider;
+import org.certis.studyplatform.shared.security.JwtAuthenticationFilter;
+import org.certis.studyplatform.shared.security.JwtTokenProvider;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
@@ -16,8 +16,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.cors.CorsConfigurationSource;
-
-import static org.springframework.security.config.Customizer.withDefaults;
 
 @Configuration
 @EnableWebSecurity
